% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{acf.dsa2}
\alias{acf.dsa2}
\title{Plot the ACF based on \code{dsa2} objects}
\usage{
\method{acf}{dsa2}(
  x,
  lags = c(1:7 + 1, 30:31 + 1, 365 + 1, 730 + 1),
  ylim = c(-1, 1),
  main = "ACF for selected lags",
  xlab = "lags",
  ylab = "ACF",
  col = .dsa2color("darkgreen"),
  col2 = .dsa2color("blue"),
  col3 = .dsa2color("darkgrey"),
  space = 6,
  border = NA,
  ...
)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{lags}{An integer value specifying which lags are shown.}

\item{ylim}{A numeric vector indicating the limits of y-axis.}

\item{main}{A character value detailing the title of the plot.}

\item{xlab}{A character value detailing the label of x-axis.}

\item{ylab}{A character value detailing the label of y-axis.}

\item{col}{A character value detailing the color of the line.}

\item{col2}{A character value detailing the color of horizontal lines for the confidence interval.}

\item{col3}{A character value detailing the color of vertical lines to separate lags of interest.}

\item{space}{A numeric value detailing the space before each bar (see \code{barplot()}).}

\item{border}{A character value detailing the color of the border.}

\item{...}{Further arguments for \code{barplot()}.}
}
\description{
Plot the ACF for a seasonally adjusted time series extracted from a \code{dsa2} object.
}
\details{
Wrapper around the \code{stats::acf}() function. See \code{barplot()} for details on changing the look of the plot
}
\examples{
x <- tssim::sim_daily(3)$original
result <- dsa(x)
acf(result)
}
\author{
Daniel Ollech
}
