% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{acf.ts}
\alias{acf.ts}
\title{ACF for \code{ts} objects}
\usage{
\method{acf}{ts}(x, ...)

\method{acf}{ts}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{parameters}
}
\description{
See \code{stats::acf} for details.

See \link[stats]{acf} for details
}
\author{
Daniel Ollech
}
\keyword{internal}
