% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{acf.wsa}
\alias{acf.wsa}
\title{ACF for \code{wsa} objects}
\usage{
\method{acf}{wsa}(x, ...)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{...}{Additional parameters.}
}
\description{
ACF applied to the outcome of a call to \code{wsa()}
}
