% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{adjust.character}
\alias{adjust.character}
\title{Character method for adjusting a seasonal time series}
\usage{
\method{adjust}{character}(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}

\item{log}{should logs be used}
}
\description{
Character method for adjusting a seasonal time series
}
\author{
Daniel Ollech
}
