% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{adjust.default}
\alias{adjust.default}
\title{Default method for adjusting a seasonal time series}
\usage{
\method{adjust}{default}(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}

\item{log}{multiplicative models used}
}
\description{
Default method for adjusting a seasonal time series
}
\author{
Daniel Ollech
}
