% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{adjust.NULL}
\alias{adjust.NULL}
\title{NULL method for adjusting a seasonal time series}
\usage{
\method{adjust}{`NULL`}(method, series)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}
}
\description{
NULL method for adjusting a seasonal time series
}
\author{
Daniel Ollech
}
