% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{adjust}
\alias{adjust}
\title{Generic for adjusting a seasonal time series}
\usage{
adjust(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}

\item{log}{should logs be used}
}
\description{
Generic for adjusting a seasonal time series
}
\author{
Daniel Ollech
}
