% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{adjust.stl_method}
\alias{adjust.stl_method}
\title{STL method for adjusting a seasonal time series}
\usage{
\method{adjust}{stl_method}(method, series)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}
}
\description{
STL method for adjusting a seasonal time series
}
\author{
Daniel Ollech
}
