% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{adjust.x11_method}
\alias{adjust.x11_method}
\title{X-11 method for adjusting a seasonal time series}
\usage{
\method{adjust}{x11_method}(method, series)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}
}
\description{
X-11 method for adjusting a seasonal time series
}
\author{
Daniel Ollech
}
