% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{compare_plot.dsa2}
\alias{compare_plot.dsa2}
\title{Function to compare two \code{dsa2} results}
\usage{
\method{compare_plot}{dsa2}(x1, x2, include_forecasts = FALSE)
}
\arguments{
\item{x1}{The first \code{dsa2} object.}

\item{x2}{The second \code{dsa2} object.}

\item{include_forecasts}{A logical value indicating whether to display forecast data.}
}
\description{
Create a plot of the original and adjusted series from two different adjustments.
}
\examples{
set.seed(2358)
all <- tssim::sim_daily(N = 5)
series  <- all$original
result  <- dsa(series, outliers = NULL)
result2 <- dsa(series, s7 = "x11", pre_processing = result)
compare_plot(result, result2)
}
\author{
Daniel Ollech
}
