% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{compare_plot}
\alias{compare_plot}
\title{Plot a comparison between two \code{dsa2} or two \code{wsa} objects}
\usage{
compare_plot(x1, x2 = NULL, include_forecasts = FALSE)
}
\arguments{
\item{x1}{The first \code{dsa2} or \code{wsa} object.}

\item{x2}{The second \code{dsa2} or \code{wsa} object.}

\item{include_forecasts}{A logical value indicating whether to display forecast data.}
}
\description{
Compare plot can be applied to \code{dsa2} and \code{wsa} objects, i.e. the results from a call to \code{dsa()} or \code{wsa()}.
}
\author{
Daniel Ollech
}
