% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{compare_plot.wsa}
\alias{compare_plot.wsa}
\title{Function to compare two \code{wsa} results}
\usage{
\method{compare_plot}{wsa}(x1, x2, include_forecasts = FALSE)
}
\arguments{
\item{x1}{First \code{wsa} object.}

\item{x2}{Second \code{wsa} object.}

\item{include_forecasts}{A logical value indicating whether to display forecast data.}
}
\description{
Create a plot of the original and adjusted series from two different
adjustments. Inspired by dsa2 functions.
}
\author{
Carolin Schaaff
}
