% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{compute_seasadj}
\alias{compute_seasadj}
\title{Compute calendar and seasonally adjusted time series}
\usage{
compute_seasadj(
  series,
  seasComp7 = NULL,
  seasComp31 = NULL,
  seasComp365 = NULL,
  calComp = NULL,
  log = TRUE
)
}
\arguments{
\item{series}{An \code{xts} object containing the time series to be adjusted.}

\item{seasComp7}{An \code{xts} object containing the day-of-the-week component.}

\item{seasComp31}{An \code{xts} object containing the day-of-the-month component.}

\item{seasComp365}{An \code{xts} object containing the day-of-the-year component.}

\item{calComp}{An \code{xts} object containing the calendar component.}

\item{log}{A logical value indicating whether to use a multiplicative model.}
}
\description{
Compute calendar and seasonally adjusted time series
}
\author{
Daniel Ollech
}
