% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{delete_29}
\alias{delete_29}
\title{Delete February 29}
\usage{
delete_29(x)
}
\arguments{
\item{x}{time series}
}
\description{
Delete the observation on February 29 from a time series
}
\author{
Daniel Ollech
}
