% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.apply_msr_rule}
\alias{.apply_msr_rule}
\title{Select an S3Xk seasonal moving-average filter}
\usage{
.apply_msr_rule(
  series,
  n_backcasts = 0,
  n_forecasts = 0,
  seasonal_periodicity = 4,
  tentative_seasonal_ma_length = 7L,
  use_additive_formula = TRUE
)
}
\arguments{
\item{series}{A numeric vector containing the time series.}

\item{n_backcasts}{An integer number of leading observations to remove
before analysis.}

\item{n_forecasts}{An integer number of trailing observations to remove
before analysis.}

\item{seasonal_periodicity}{A numeric or integer vector of length-1 specifying
the seasonal periodicity.}

\item{tentative_seasonal_ma_length}{An odd integer greater than 1 that
gives the length of the simple moving average used for the tentative seasonal
estimate (default 7).}

\item{use_additive_formula}{A logical value whether to use additive decomposition (\code{TRUE};
irregular = series - seasonal) or multiplicative decomposition (\code{FALSE};
irregular = series / seasonal).}
}
\value{
a list containing the selected seasonal_filter and the global MSR
}
\description{
Determine the appropriate S3Xk seasonal moving-average filter ("S3X1", "S3X3",
"S3X5", "S3X9", "S3X15", or "S3X23") for a given time series by computing a
global MSR statistic from tentative seasonal and irregular components. The
procedure can operate in additive or multiplicative mode.
}
\details{
The method:
\itemize{
\item Optionally removes user-specified backcasts and forecasts.
\item Trims the series to an integer multiple of the minimum seasonal periodicity.
\item Obtains tentative seasonal and irregular components via moving averages.
\item Computes a global MSR statistic and rescales it using a factor following
Lothian (1978).
\item Selects the S3Xk filter based on intersection-rule thresholds.
}

If no filter is selected at the current length, the series is iteratively
shortened by one minimum seasonal cycle and re-evaluated. If no selection is
possible after shortening or the series was shortened 100 times, the function
returns "S3X5" by default.
\itemize{
\item The function requires at least five seasonal cycles to evaluate the thresholds;
otherwise, it progressively trims the series.
\item Input validation ensures a feasible seasonal configuration and appropriate
moving-average length.
}
}
\references{
Lothian, J. (1978), The Identification and Treatment of Moving Seasonality
in the X-11 Seasonal Adjustment Method, Research Paper 78-10-004, Seasonal
Adjustment and Time Series Analysis Staff, Statistics Canada.
}
\author{
Karsten Webel, Lea Hengen, Daniel Ollech
}
\keyword{internal}
