% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{.correct_filter_length}
\alias{.correct_filter_length}
\title{Check the chosen X-11 filter}
\usage{
.correct_filter_length(method, series)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted, class should be ts}
}
\description{
If X-11 is used on the day-of-the-year effect, the number of observations included in the time series needs to be high enough for certain filters to be used. This function checks it.
}
\author{
Daniel Ollech
}
