% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.descaler}
\alias{.descaler}
\title{Internal function to invert taking logs and differences of a time series}
\usage{
.descaler(x, y = NA, Diff = 0, log = FALSE, Lag = NA)
}
\arguments{
\item{x}{An \code{xts} object containing the time series to be adjusted.}

\item{y}{An \code{xts} object containing the reference time series for inverting differencing.}

\item{Diff}{An integer value detailing the number of differences to be taken.}

\item{log}{A logical value indicating whether the time series should be logarithmised.}

\item{Lag}{An integer value detailing which lag of differencing needs to be inverted.}
}
\description{
For a series that has been logged and/or differenced, this function reverses these transformations.
}
\details{
The function is used in \code{dsa()} to handle the users choice of logs and levels.
}
\author{
Daniel Ollech
}
\keyword{internal}
