% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.dsa2color}
\alias{.dsa2color}
\title{Defining dsa2 colors}
\usage{
.dsa2color(color, ...)
}
\arguments{
\item{color}{A character value specifying the name of the color.}

\item{...}{Additional colors}
}
\description{
This function defines the colors to be used throughout dsa2.
}
\author{
Daniel Ollech
}
\keyword{internal}
