% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{.estimate_component.character}
\alias{.estimate_component.character}
\title{Character method for estimating seasonal component}
\usage{
\method{.estimate_component}{character}(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted, class should be ts}

\item{log}{should logs be used}
}
\description{
Character method for estimating seasonal component
}
\author{
Daniel Ollech
}
