% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{.estimate_component.stl_method}
\alias{.estimate_component.stl_method}
\title{STL method for estimating seasonal component}
\usage{
\method{.estimate_component}{stl_method}(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted, class should be ts}

\item{log}{should logs be used}
}
\description{
STL method for estimating seasonal component
}
\author{
Daniel Ollech
}
\keyword{internal}
