% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.msr_rule}
\alias{.msr_rule}
\title{Translate msr value into a filter}
\usage{
.msr_rule(global_msr)
}
\arguments{
\item{global_msr}{A numeric containing the global msr value of a time series.}
}
\description{
Translate msr value into a filter.
}
\author{
Karsten Webel, Lea Hengen, Daniel Ollech
}
\keyword{internal}
