% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.msr_rule_ultimate}
\alias{.msr_rule_ultimate}
\title{Translate msr value into filter even if it repeatedly falls in between thresholds}
\usage{
.msr_rule_ultimate(global_msr)
}
\arguments{
\item{global_msr}{A numeric containing the global msr value of a time series}
}
\description{
Translate msr value into filter
}
\author{
Daniel Ollech, Lea Hengen
}
\keyword{internal}
