% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Functions to be run when package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{A character value containing the name of the library.}

\item{pkgname}{A character value containing the name of the package.}
}
\description{
Functions to be run when package is loaded.
}
\keyword{internal}
