% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{.outARIMA.wsa}
\alias{.outARIMA.wsa}
\title{Internal function for Fractional Airline model}
\usage{
\method{.outARIMA}{wsa}(x, digits = 3)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{digits}{An integer value specifying the number of digits to be displayed.}
}
\description{
Internal function to polish the output for the fractional airline model.
Inspired by dsa2 functions.
}
\author{
Carolin Schaaff
}
\keyword{internal}
