% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{.outCalendar.wsa}
\alias{.outCalendar.wsa}
\title{Internal function for calendars}
\usage{
\method{.outCalendar}{wsa}(x, digits = 3)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{digits}{An integer value specifying the number of digits to be displayed.}
}
\description{
Internal function to polish the output for calendars.
Inspired by dsa2 functions.
}
\author{
Carolin Schaaff
}
\keyword{internal}
