% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.outOutlier.dsa2}
\alias{.outOutlier.dsa2}
\title{Internal function for outliers}
\usage{
\method{.outOutlier}{dsa2}(x, digits = 3)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{digits}{An integer value specifying the number of digits to be displayed.}
}
\description{
Internal function to polish the output for outliers.
}
\author{
Sindy Brakemeier, Lea Hengen
}
\keyword{internal}
