% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.outOutlier}
\alias{.outOutlier}
\title{Display outlier results for \code{dsa2} and \code{wsa} objects}
\usage{
.outOutlier(x, digits = 3)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{digits}{An integer value specifying the number of digits to be displayed.}
}
\description{
Display outlier results for \code{dsa2} and \code{wsa} objects.
}
\author{
Daniel Ollech
}
\keyword{internal}
