% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{.outOutlier.wsa}
\alias{.outOutlier.wsa}
\title{Internal function for outliers}
\usage{
\method{.outOutlier}{wsa}(x, digits = 3)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{digits}{An integer value specifying the number of digits to be displayed.}
}
\description{
Internal function to polish the output for outliers.
Inspired by dsa2 functions.
}
\author{
Carolin Schaaff
}
\keyword{internal}
