% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.rescaling_factor_lothian}
\alias{.rescaling_factor_lothian}
\title{Helper function for MSR selection}
\usage{
.rescaling_factor_lothian(n_observations = 10L)
}
\arguments{
\item{n_observations}{An integer detailing the number of observations.}
}
\description{
Local helper: rescaling factors (Lothian, 1978).
}
\author{
Karsten Webel, Lea Hengen, Daniel Ollech
}
\keyword{internal}
