% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.run_msr}
\alias{.run_msr}
\title{Use msr}
\usage{
.run_msr(series, parameters)
}
\arguments{
\item{series}{series to be adjusted}

\item{parameters}{parameters for x11plus}
}
\description{
Applies new msr rule
}
\details{
Calculates a preliminary d8 assuming weekly and yearly seasonality and applies the new msr rule. Returns the chosen seasonal moving average
}
\author{
Lea Hengen, Daniel Ollech
}
\keyword{internal}
