% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.scaler}
\alias{.scaler}
\title{Internal function to take logs and differences of a time series}
\usage{
.scaler(x, Diff = 0, log = FALSE)
}
\arguments{
\item{x}{An \code{xts} object containing the time series to be adjusted.}

\item{Diff}{An integer value detailing the number of differences to be taken.}

\item{log}{A logical value indicating whether the time series should be logarithmised.}
}
\description{
Logarithmise and / or difference a time series
}
\details{
The function is used in \code{dsa()} to handle the users choice of logs and levels.
}
\author{
Daniel Ollech
}
\keyword{internal}
