% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.single_plot_spectrum}
\alias{.single_plot_spectrum}
\title{Helper function for the plotting of spectra}
\usage{
.single_plot_spectrum(df, ylab = "Spectrum", title = "Spectrum", dsa2 = TRUE)
}
\arguments{
\item{df}{A data frame with spectrum data.}

\item{ylab}{A character value detailing the label of y-axis.}

\item{title}{A character value detailing the title of the plot.}

\item{dsa2}{A logical value indicating whether the object in \code{spectrum()} is a \code{dsa2} object.}
}
\description{
The function is used in \code{spectrum.wsa()} and \code{spectrum.dsa2()}.
}
\author{
Daniel Ollech
}
\keyword{internal}
