% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.table}
\alias{.table}
\title{Output table for \code{dsa2} objects}
\usage{
.table(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Additional parameters for \code{gt::tab_options}.}
}
\description{
Creates a table to be used in \code{output()}.
}
\details{
The function uses the gt package.
}
\author{
Martin Stefan, Daniel Ollech
}
\keyword{internal}
