% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{dsa2}
\alias{dsa2}
\title{Seasonal adjustment of daily time series}
\usage{
dsa2(
  series,
  xreg = NULL,
  log = TRUE,
  s7 = c("x11", "stl", "seats")[2],
  s31 = NULL,
  s365 = c("x11", "stl", "seats")[2],
  outliers = c("AO", "LS", "TC"),
  n_iterations = 1,
  h = 365,
  interpolator = "CUBIC_SPLINE",
  pre_processing = NULL,
  ...
)
}
\arguments{
\item{series}{input time series in xts format}

\item{xreg}{regressors used for calendar adjustment}

\item{log}{should multiplicative time series model be used}

\item{s7}{method or specification used for adjustment of day-of-the-week}

\item{s31}{method or specification used for adjustment of day-of-the-month}

\item{s365}{method or specification used for adjustment of day-of-the-year}

\item{outliers}{should outliers be identified using regarima models}

\item{n_iterations}{number of iterations of step 2 to 4 (i.e. s7, s31 and s365)}

\item{h}{number of days to forecast}

\item{interpolator}{either "CUBIC_SPLINE" or "NONE" (inherited from rjd3bbksplines::interpolate31)}

\item{pre_processing}{Optionally include pre-processing results computed earlier}

\item{...}{additional parameters from fractionalAirlineEstimation}
}
\description{
Seasonal adjustment of daily time series
}
\details{
This function implements the DSA2 procedure that facilitates the
seasonal adjustment of daily time series.
}
\examples{
## Create time series 
set.seed(2358)
all <- tssim::sim_daily(N=5)
series <- all$original 

## Default adjustment
result <- dsa2(series)
plot(result)

## Set STL parameters to be used
result2 <- dsa2(series, s7 = stl_method(swindow = 31, twindow=9))

## Use STL and X11 in combination
result3 <- dsa2(series, s7 = "stl", s365 = "x11")

## Compare results
compare_plot(result2, result3)
}
\references{
Ollech, Daniel (2018). Seasonal Adjustment of Daily Time Series.
Bundesbank Discussion Paper 41/2018.

Ollech, Daniel (2021). Seasonal Adjustment of Daily Time Series.
Journal of Time Series Econometrics 13 (2), 235-264.
}
\author{
Daniel Ollech, Martin Stefan
}
