% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{estimate_component.default}
\alias{estimate_component.default}
\title{Default method for estimating seasonal component}
\usage{
\method{estimate_component}{default}(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}

\item{log}{multiplicative models used}
}
\description{
Default method for estimating seasonal component
}
\author{
Daniel Ollech
}
