% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{estimate_component}
\alias{estimate_component}
\title{Generic for estimating seasonal component}
\usage{
estimate_component(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}

\item{log}{should logs be used}
}
\description{
Generic for estimating seasonal component
}
\author{
Daniel Ollech
}
