% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{estimate_component.seats_method}
\alias{estimate_component.seats_method}
\title{Seats method for estimating seasonal component}
\usage{
\method{estimate_component}{seats_method}(method, series, log = NULL)
}
\arguments{
\item{method}{method to be employed}

\item{series}{time series to be adjusted}

\item{log}{should logs be used}
}
\description{
Seats method for estimating seasonal component
}
\author{
Daniel Ollech
}
