% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{holidaysCentered}
\alias{holidaysCentered}
\title{Centered version of xts of all German regressors used for daily time series}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 55152 rows and 132 columns.
}
\usage{
holidaysCentered
}
\description{
This is the centered version of the holidays regressors. See ?dsa2::holidays for details
}
\author{
Daniel Ollech
}
\keyword{datasets}
