% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{pacf.dsa2}
\alias{pacf.dsa2}
\title{Plot the PACF based on \code{dsa2} objects}
\usage{
\method{pacf}{dsa2}(
  x,
  lag.max = 730,
  plot = TRUE,
  na.action = stats::na.exclude,
  lags = c(1:7, 30:31, 365, 730),
  ylim = c(-1, 1),
  main = "PACF for selected lags",
  xlab = "lags",
  ylab = "PACF",
  col = .dsa2color("darkgreen"),
  col2 = .dsa2color("blue"),
  col3 = .dsa2color("darkgrey"),
  space = 6,
  border = NA,
  ...
)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{lag.max}{A integer value specifying the maximum lags, included for consistency with \code{pacf.default()}.}

\item{plot}{A logical value indicating whether the object is plotted, included for consistency with \code{pacf.default()}.}

\item{na.action}{see \code{stats::na.exclude}, handling of missing observations.}

\item{lags}{An integer value specifying which lags are shown.}

\item{ylim}{A numeric vector indicating the limits of y-axis.}

\item{main}{A character value detailing the title of the plot.}

\item{xlab}{A character value detailing the label of x-axis.}

\item{ylab}{A character value detailing the label of y-axis.}

\item{col}{A character value detailing the color of the line.}

\item{col2}{A character value detailing the color of horizontal lines for the confidence interval.}

\item{col3}{A character value detailing the color of vertical lines to separate lags of interest.}

\item{space}{A numeric value detailing the space before each bar (see \code{barplot()}).}

\item{border}{A character value detailing the color of the border.}

\item{...}{Further arguments for \code{barplot()}.}
}
\description{
Plot the PACF for a seasonally adjusted time series extracted from a \code{dsa2} object.
}
\details{
Wrapper around the \code{stats::pacf()} function. See \code{barplot()} for details on changing the look of the plot.
}
\examples{
x <- tssim::sim_daily(3)$original
result <- dsa(x)
pacf(result)
}
\author{
Daniel Ollech
}
