% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{pacf.numeric}
\alias{pacf.numeric}
\title{PACF for numeric object}
\usage{
\method{pacf}{numeric}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{parameters}
}
\description{
See ?stats::pacf for details
}
\keyword{internal}
