% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{pacf.wsa}
\alias{pacf.wsa}
\title{PACF for \code{wsa} objects}
\usage{
\method{pacf}{wsa}(x, ...)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{...}{Additional parameters.}
}
\description{
PACF applied to the outcome of a call to \code{wsa()}.
}
\keyword{internal}
