% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot.dsa2}
\alias{plot.dsa2}
\title{Plot the results of \code{dsa2} objects}
\usage{
\method{plot}{dsa2}(
  x,
  main = "Result for seasonal adjustment of daily time series",
  include_forecasts = FALSE,
  type = "l",
  xlab = "",
  ylab = "",
  cex.axis = 0.75,
  bty = "n",
  mar = c(4, 2, 2, 0.5),
  ...
)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{main}{A character value detailing the title of the plot.}

\item{include_forecasts}{A logical value indicating whether to display forecast data.}

\item{type}{A single character indicating the line type (see \code{plot()}).}

\item{xlab}{A character value detailing the label of x-axis.}

\item{ylab}{A character value detailing the label of y-axis (possibly change the margins using parameter \code{mar}).}

\item{cex.axis}{see \code{par()}.}

\item{bty}{see \code{par()}.}

\item{mar}{A numerical vector indicating the number of lines of margin to be specified on the four sides of the plot (see \code{par()}).}

\item{...}{Additional parameters from \code{plot()}}
}
\description{
Creates a plot of original and seasonally adjusted series.
}
\details{
The function uses the base plot package.
}
\author{
Sindy Brakemeier, Daniel Ollech
}
