% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{plot.wsa}
\alias{plot.wsa}
\title{Plot for \code{wsa} objects}
\usage{
\method{plot}{wsa}(
  x,
  main = "Result for seasonal adjustment of weekly time series",
  include_forecasts = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{main}{A character value detailing the title of the plot.}

\item{include_forecasts}{A logical value indicating whether to display forecast data.}

\item{...}{Additional parameters from \code{plot()} function}
}
\description{
Creates a plot of original and seasonally adjusted series.
Aligned with dsa2 functions.
}
\details{
The function uses the base plot package.
}
\author{
Carolin Schaaff
}
