% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Interactive plots}
\usage{
plot_interactive(x, ...)
}
\arguments{
\item{x}{An \code{dsa2}, \code{wsa} or \code{xts} object to be plotted.}

\item{...}{Additonal parameters.}
}
\description{
The interactive plot can be applied to a \code{dsa2}, \code{wsa} and \code{xts} object. Is it based on dygraphs.
}
\author{
Daniel Ollech
}
