% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot_interactive.wsa}
\alias{plot_interactive.wsa}
\title{Interactive plot for \code{wsa} objects}
\usage{
\method{plot_interactive}{wsa}(x, ...)
}
\arguments{
\item{x}{An \code{wsa} object to be plotted.}

\item{...}{Additional parameters for \code{dygraphs::dyOptions}.}
}
\description{
Creates a plot of original and seasonally adjusted series.
}
\details{
The function uses the dygraphs package.
}
\author{
Daniel Ollech
}
