% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{print.dsa2}
\alias{print.dsa2}
\title{Print for \code{dsa2} objects}
\usage{
\method{print}{dsa2}(x, ...)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{...}{Further arguments to print.}
}
\description{
Print for \code{dsa2} objects, lists the coefficients of the fractional airline model as well as the coefficients and
t-values of all outliers and calendar effects.
}
\author{
Sindy Brakemeier, Lea Hengen
}
\keyword{internal}
