% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{print.wsa}
\alias{print.wsa}
\title{Print for \code{wsa} objects}
\usage{
\method{print}{wsa}(x, ...)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{...}{Further arguments to print.}
}
\description{
Lists the coefficients of the fractional airline model
as well as the coefficients and t-values of all outliers and calendar effects.
Inspired by dsa2 functions.
}
\author{
Carolin Schaaff
}
