% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{run_msr}
\alias{run_msr}
\title{Use msr}
\usage{
run_msr(series, parameters, period = NULL)
}
\arguments{
\item{series}{An \code{xts} object containing the time series to be adjusted.}

\item{parameters}{A list of parameters for x11plus (defined in \code{x11_method()}).}

\item{period}{A numeric value specifying the frequency to be adjusted.}
}
\description{
Applies new msr rule
}
\details{
Calculates a preliminary D8 assuming weekly and yearly seasonality and applies the new msr rule.
Returns the chosen seasonal moving average.
}
\author{
Lea Hengen, Daniel Ollech
}
