% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{run_msr_weekly}
\alias{run_msr_weekly}
\title{Use msr for a weekly time series}
\usage{
run_msr_weekly(series, parameters, period = NULL)
}
\arguments{
\item{series}{series to be adjusted}

\item{parameters}{parameters for x11plus}

\item{period}{period of time series}
}
\description{
Applies new msr rule
}
\details{
Returns the chosen seasonal moving average
}
\author{
Lea Hengen, Daniel Ollech
}
