% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{scaler}
\alias{scaler}
\title{Internal function to take logs and differences of  a time series}
\usage{
scaler(x, Diff = 0, log = FALSE)
}
\arguments{
\item{x}{time series}

\item{Diff}{number of differences to be taken}

\item{log}{Should time series be logarithmised}
}
\description{
Logarithmise and / or difference a time series
}
\details{
Function is used in dsa to handle the users choice of logs and levels.
}
\author{
Daniel Ollech
}
