% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{seats_method}
\alias{seats_method}
\title{Translation between input to \code{dsa()} and rjd3-function}
\usage{
seats_method(
  period = NA,
  log = NULL,
  sn = FALSE,
  stde = FALSE,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{period}{A numeric value specifying the frequency to be adjusted.}

\item{log}{A logical value choosing whether to use a multiplicative model.}

\item{sn}{A logical value choosing whether the series should be decomposed into two components (signal and noise).}

\item{stde}{A logical value choosing whether to compute standard deviations of components.}

\item{nbcasts}{A numeric scalar detailing the number of forecasts.}

\item{nfcasts}{A numeric scalar detailing the number of backcasts.}
}
\description{
Handler for Seats
}
\details{
This functions is basically a translator between the dsa2 routines
and \code{rjd3highfreq::fractionalAirlineDecomposition}, but its goal is to invoke the Seats procedure.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in \code{dsa()}.
}
\author{
Daniel Ollech
}
