% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{spectrum.dsa2}
\alias{spectrum.dsa2}
\title{Plot the periodogram for \code{dsa2} objects}
\usage{
\method{spectrum}{dsa2}(x, ...)
}
\arguments{
\item{x}{A \code{dsa2} object.}

\item{...}{Further options to \code{par()}.}
}
\description{
Plot the periodogram of the adjustment result of a daily time series.
}
\details{
The spectrum is build around the \code{spec.pgram()} function.
}
\examples{
x <- tssim::sim_daily(3)$original
res <- dsa(x)
spectrum(res)
}
\author{
Daniel Ollech
}
