% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{spectrum.numeric}
\alias{spectrum.numeric}
\title{Spectrum for numeric objects}
\usage{
\method{spectrum}{numeric}(x, ...)

\method{spectrum}{numeric}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{parameters}
}
\description{
See \code{stats::spectrum} for details.

See \link[stats]{spectrum} for details
}
\author{
Daniel Ollech
}
\keyword{internal}
