% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa_auxiliary.R
\name{spectrum.wsa}
\alias{spectrum.wsa}
\title{Plot the periodogram for \code{wsa} objects}
\usage{
\method{spectrum}{wsa}(x, ...)
}
\arguments{
\item{x}{A \code{wsa} object.}

\item{...}{Further options to \code{par()}.}
}
\description{
Plot the periodogram of the adjustment result of a weekly time series.
}
\details{
The spectrum is build around the \code{spec.pgram()} function
}
\author{
Daniel Ollech
}
