% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{stats_stl_method}
\alias{stats_stl_method}
\title{Translation between input to \code{dsa()} and \code{stats::stl()}-function}
\usage{
stats_stl_method(
  swindow = 13,
  twindow = NA,
  lwindow = NA,
  ninnerloop = 1,
  nouterloop = 15,
  sjump = NA,
  tjump = NA,
  ljump = NA
)
}
\arguments{
\item{swindow}{A numeric value for the number of observations included in the local regressions calculated to obtain the seasonal component.}

\item{twindow}{A numeric value for the number of observations included in local regressions for trend component.}

\item{lwindow}{A numeric value for the number of observations included in local regressions for additional trend smoothing of seasonal component.}

\item{ninnerloop}{A numeric value for the number of inner loops of STL.}

\item{nouterloop}{A numeric value for the number of outer loops of STL.}

\item{sjump}{An integer value to set "jump speed" for the linear interpolation for the seasonal component.
Linear interpolation happens between every *.jump-th value (see \code{stats::stl()}).}

\item{tjump}{An integer value to set "jump speed" for the linear interpolation for the trend component (see \code{stats::stl()}).#
Linear interpolation happens between every *.jump-th value (see \code{stats::stl()}).}

\item{ljump}{An integer value to set "jump speed" for the linear interpolation for the seasonal subseries component (see \code{stats::stl()}).
Linear interpolation happens between every *.jump-th value (see \code{stats::stl()}).}
}
\description{
Handler for \code{stats::stl()}
}
\details{
This functions is basically a translator between the dsa2 routines and \code{stats::stl()}.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in \code{dsa()}.
}
\author{
Daniel Ollech
}
