% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{stl_method}
\alias{stl_method}
\title{Translation between input to \code{dsa()} and rjd3-function}
\usage{
stl_method(
  period = NA,
  swindow = 13,
  log = NULL,
  twindow = 0,
  lwindow = 0,
  ninnerloop = 1,
  nouterloop = 15,
  sjump = 0,
  tjump = 0,
  ljump = 0,
  weight.threshold = 0.001,
  weight.function = c("BIWEIGHT"),
  legacy = FALSE
)
}
\arguments{
\item{period}{A numeric value specifying the frequency to be adjusted.}

\item{swindow}{Number of observations included in the local regressions calculated to obtain the seasonal component.}

\item{log}{A logical value indicating whether to use a multiplicative model., ignored in dsa2. ## MR2: How is it ignored??}

\item{twindow}{A numeric value for the number of observations included in local regressions for trend component.}

\item{lwindow}{A numeric value for the number of observations included in local regressions for additional trend smoothing of seasonal component.}

\item{ninnerloop}{A numeric value for the number of inner loops of STL.}

\item{nouterloop}{A numeric value for the number of outer loops of STL.}

\item{sjump}{A numeric value for the number of jumps in the computation of the seasonal component.}

\item{tjump}{A numeric value for the number of jumps in the computation of the trend component.}

\item{ljump}{A numeric value for the number of jumps in the computation of the trend in the seasonal component.}

\item{weight.threshold}{A numeric value detailing the threshold for weights (see \code{?rjd3stl::stlplus}).}

\item{weight.function}{Function for weights (see \code{?rjd3stl::stlplus}).}

\item{legacy}{Use of the legacy MAD.}
}
\description{
Handler for STL
}
\details{
This functions is basically a translator between the dsa2 routines and \code{rjd3stl::stlplus},
but its goal is to invoke the STL-procedure. It cannot be used to change the decomposition scheme
(additive/multiplicative) for a single step in \code{dsa()}.
}
\author{
Daniel Ollech
}
