% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{wsa}
\alias{wsa}
\title{Seasonal adjustment of weekly time series}
\usage{
wsa(
  x,
  sa_method = c("seats", "x11")[1],
  reiterate = 1,
  xreg = NULL,
  log = TRUE,
  outliers = c("AO", "LS", "WO"),
  h = 53,
  critical_value = 5,
  ...
)
}
\arguments{
\item{x}{An \code{xts} object containing the time series to be adjusted.}

\item{sa_method}{A character value containing either "x11" or "seats" or a call to x11_method() or seats_method().}

\item{reiterate}{An integer value detailing the number of reiterations for the seasonal adjustment.}

\item{xreg}{Regressors used for calendar adjustment.}

\item{log}{A logical value indicating whether to use a multiplicative model.}

\item{outliers}{Identifying outliers (LS - leve shift, AO - additive outlier, WO - ???); set \code{NULL} if none shall be searched.}

\item{h}{A numeric value specifying the number of days to forecast.}

\item{critical_value}{A numerical value specifying the threshold to include outliers used in the automatic outlier estimation (see also
\code{rjd3highfreq::fractionalAirlineEstimation}).}

\item{...}{Additional parameters from \code{rjd3highfreq::fractionalAirlineEstimation}.}
}
\value{
An object of class "wsa", a list with elements:
series: xts with columns original and seas_adj
components: xts with calendar and seasonal components
preProcessing: model fit from rjd3highfreq::fractionalAirlineEstimation
decomposition: seasonal adjustment details from the chosen method
specification: a list of settings (e.g., log)
parameters: call parameters and metadata
}
\description{
Seasonal adjustment of weekly time series.
}
\details{
This function can be used to seasonally and calendar adjust weekly time series. Currently the function uses the extended airline estimation model from the rjd3highfreq package for calendar estimation and the extended airline decomposition model from the same package for seasonality estimation.
}
\examples{
\dontrun{
to_week <- function (x) {
ep <- xts::endpoints(x, on = "weeks")
out <- xts::period.apply(x, 
                         INDEX = ep[ep >= 0 &  ep <= length(x)], 
                         FUN = mean)
return(out)
}
seriesw <- to_week(series)
res <- wsa(seriesw)
plot(res)
}
}
\author{
Daniel Ollech, Carolin Schaaff
}
