% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{x11_method}
\alias{x11_method}
\title{Translation between input to \code{dsa()} and rjd3-function}
\usage{
x11_method(
  period = NA,
  log = NULL,
  sma = c("MSR", "S3X9", "S3X1", "S3X3", "S3X5", "S3X15")[1],
  trend.horizon = NULL,
  trend.degree = NULL,
  trend.kernel = c("Henderson"),
  trend.asymmetric = c("CutAndNormalize"),
  sigma = c(1.5, 2.5)
)
}
\arguments{
\item{period}{A numeric value specifying the frequency to be adjusted.}

\item{log}{A logical value indicating whether to use a multiplicative model.}

\item{sma}{Seasonal moving average used.}

\item{trend.horizon}{Band width of trend filters (see \code{?rjd3x11plus::x11plus}).}

\item{trend.degree}{Polynomial order of the local trend model (see \code{?rjd3x11plus::x11plus}).}

\item{trend.kernel}{Kernel weights in objective function (see \code{?rjd3x11plus::x11plus}).}

\item{trend.asymmetric}{Truncation type to obtain asymmetric from symmetric filter (see \code{?rjd3x11plus::x11plus}).}

\item{sigma}{Lower and upper limit for the extreme value detection.}
}
\description{
Handler for x-11
}
\details{
This functions is basically a translator between the dsa2 routines
and \code{rjd3x11plus::x11plus}, but its goal is to invoke the X-11 procedure.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in \code{dsa()}.
}
\author{
Daniel Ollech
}
