% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{acf.dsa2}
\alias{acf.dsa2}
\title{Plot the ACF based on dsa2 object}
\usage{
\method{acf}{dsa2}(
  x,
  lags = c(1:7 + 1, 30:31 + 1, 365 + 1, 730 + 1),
  ylim = c(-1, 1),
  main = "ACF for selected lags",
  xlab = "lags",
  ylab = "ACF",
  col = .dsa2color("darkgreen"),
  col2 = .dsa2color("blue"),
  col3 = .dsa2color("darkgrey"),
  space = 6,
  border = NA,
  ...
)
}
\arguments{
\item{x}{object as calculated by dsa2()}

\item{lags}{which lags shall be shown}

\item{ylim}{limits of y-axis}

\item{main}{title of plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{col}{color of lines}

\item{col2}{color of horizontal lines for confidence interval}

\item{col3}{color of vertical lines to separate lags of interest}

\item{space}{space before each bar}

\item{border}{color of border}

\item{...}{further arguments for barplot()}
}
\description{
Plot the ACF for a seasonally adjusted time series extracted from a dsa2 object
}
\details{
Wrapper around the stats::acf() function. See ?barplot for details
on changing the look of the plot
}
\examples{
x <- tssim::sim_daily(3)$original
result <- dsa(x)
acf(result)
}
\author{
Daniel Ollech
}
