% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{acf.dsa2}
\alias{acf.dsa2}
\title{Plot the ACF based on dsa2 object}
\usage{
\method{acf}{dsa2}(dsa2_object, ...)
}
\arguments{
\item{dsa2_object}{object as calculated by dsa2()}

\item{...}{further arguments for barplot()}
}
\description{
Plot the ACF for a seasonally adjusted time series extracted from a dsa2 object
}
\details{
Wrapper around the stats::acf() function
}
\examples{
x <- tssim::sim_daily(3)$original
result <- dsa(x)
acf(result)
}
\author{
Daniel Ollech
}
