% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{acf.numeric}
\alias{acf.numeric}
\title{ACF for numeric object}
\usage{
\method{acf}{numeric}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{parameters}
}
\description{
See ?stats::acf for details
}
