% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{acf}
\alias{acf}
\title{Generic for ACF}
\usage{
acf(x, ...)
}
\arguments{
\item{x}{object for which ACF shall be plotted}

\item{...}{parameters}
}
\description{
Generic for ACF
}
\author{
Daniel Ollech
}
