% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{compare_plot}
\alias{compare_plot}
\title{Function to compare two dsa2 results}
\usage{
compare_plot(dsa2_object1, dsa2_object2, include_forecasts = FALSE)
}
\arguments{
\item{dsa2_object1}{first dsa2 output object}

\item{dsa2_object2}{second dsa2 output object}
}
\description{
Create a plot of the original and adjusted series from two different adjustments
}
\author{
Daniel Ollech
}
