% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{compute_seasadj}
\alias{compute_seasadj}
\title{Compute calendar and seasonally adjusted time series}
\usage{
compute_seasadj(
  series,
  seasComp7 = NULL,
  seasComp31 = NULL,
  seasComp365 = NULL,
  calComp = NULL,
  log = TRUE
)
}
\arguments{
\item{series}{original, xts time series}

\item{seasComp7}{day-of-the-week component, xts time series}

\item{seasComp31}{day-of-the-month component, xts time series}

\item{seasComp365}{day-of-the-year component, xts time series}

\item{calComp}{calendar component, xts time series}

\item{log}{do we use a multiplicative model}
}
\description{
Compute calendar and seasonally adjusted time series
}
\details{
All time series used should be of class xts or a numeric constant.
}
\author{
Daniel Ollech
}
