% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{compute_seasadj}
\alias{compute_seasadj}
\title{Compute calendar and seasonally adjusted time series}
\usage{
compute_seasadj(
  xLinear,
  seasComp7,
  seasComp31,
  seasComp365,
  calComp = NULL,
  log = TRUE
)
}
\arguments{
\item{seasComp7}{day-of-the-week component}

\item{seasComp31}{day-of-the-month component}

\item{seasComp365}{day-of-the-year component}

\item{calComp}{calendar component}

\item{xLinx}{basic series}
}
\description{
Compute calendar and seasonally adjusted time series
}
\author{
Daniel Ollech
}
