% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.descaler}
\alias{.descaler}
\title{Internal function to invert taking logs and differences of a time series}
\usage{
.descaler(x, y = NA, Diff = 0, log = FALSE, Lag = NA)
}
\arguments{
\item{x}{time series}

\item{y}{reference time series for inverting differencing}

\item{Diff}{number of differences to be taken}

\item{log}{should time series be logarithmised}

\item{Lag}{which Lag of differencing needs to be inverted}
}
\description{
For a series that has been logged and/or differenced, this function reverses these transformations.
}
\details{
Function is used in dsa to handle the users choice of logs and levels.
}
\author{
Daniel Ollech
}
\keyword{internal}
