% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.dsa2color}
\alias{.dsa2color}
\title{Defining dsa2 colors}
\usage{
.dsa2color(color, ...)
}
\arguments{
\item{color}{name of color}

\item{...}{additional colors}
}
\description{
This function defines the colors to be used troughout the dsa2 package
}
\examples{
dsa2:::.dsa2color("darkblue")
}
\author{
Daniel Ollech
}
