% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_startfunction.R
\name{.onLoad}
\alias{.onLoad}
\title{Functions to be run when package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\description{
Functions to be run when package is loaded
}
\author{
Daniel Ollech
}
