% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.outARIMA}
\alias{.outARIMA}
\title{Internal function for Fractional Airline model}
\usage{
.outARIMA(dsa2_object, digits = 3)
}
\arguments{
\item{dsa2_object}{dsa2 output object}
}
\description{
Internal function to polish the output for the fractional airline model
}
\author{
Jakob Oberhammer, Martin Stefan
}
