% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.outCalendar}
\alias{.outCalendar}
\title{Internal function for calendars}
\usage{
.outCalendar(dsa2_object, digits = 3)
}
\arguments{
\item{dsa2_object}{dsa2 output object}
}
\description{
Internal function to polish the output for calendars
}
\author{
Sindy Brakemeier, Lea Hengen
}
