% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.outOutlier}
\alias{.outOutlier}
\title{Internal function for outliers}
\usage{
.outOutlier(dsa2_object)
}
\arguments{
\item{dsa2_object}{dsa2 output object}
}
\description{
Internal function to polish the output for outliers
}
\author{
Sindy Brakemeier, Lea Hengen
}
