% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{.table}
\alias{.table}
\title{Output table for dsa2}
\usage{
.table(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{additional parameters for gt::tab_options}
}
\description{
Creates a table to be used in output()
}
\details{
The function uses the gt package.
}
\author{
Martin Stefan, Daniel Ollech
}
\keyword{internal}
