% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{dsa}
\alias{dsa}
\title{Seasonal adjustment of daily time series}
\usage{
dsa(
  series,
  xreg = NULL,
  log = TRUE,
  s7 = c("x11", "stl", "seats")[2],
  s31 = NULL,
  s365 = c("x11", "stl", "seats")[2],
  outliers = c("AO", "LS", "WO"),
  critical_value = 6,
  n_iterations = 1,
  h = 365,
  interpolator = "default",
  pre_processing = NULL,
  ...
)
}
\arguments{
\item{series}{input time series in xts format}

\item{xreg}{regressors used for calendar adjustment}

\item{log}{should multiplicative time series model be used}

\item{s7}{method or specification used for adjustment of day-of-the-week}

\item{s31}{method or specification used for adjustment of day-of-the-month}

\item{s365}{method or specification used for adjustment of day-of-the-year}

\item{outliers}{which outliers should be identified (LS, AO, WO). Set NULL if none shall be searched}

\item{critical_value}{threshold to include outliers to be used in the automatic outlier estimation}

\item{n_iterations}{number of iterations of step 2 to 4 (i.e. s7, s31 and s365), at least 1}

\item{h}{number of days to forecast}

\item{interpolator}{either "default" ("NONE" if stl is used, "CUBIC_SPLINES" else), "CUBIC_SPLINE" or "NONE". See details}

\item{pre_processing}{Optionally include pre-processing results computed earlier using dsa2 result (see examples)}

\item{...}{additional parameters from fractionalAirlineEstimation}
}
\description{
Seasonal adjustment of daily time series
}
\details{
DSA iteratively estimates and adjusts the calendar component, the day-of-the-week effect, if selected: the day-of-the-month effect, and the day-of-the-year effect to get the seasonally adjusted series.
For the estimation of the day-of-the-month effect, the months are extended to include 31 days in each months. This is done by filling up the artificial days (e.g. 31st of April) by NAs and then - if so chosen - filling up the missing values using spline interpolation. By default, if STL is used, the NAs are not filled up, if X-11 or Seats is used- the NAs are filled up.
Is is not possible to use different decomposition schemes for the single dsa2 steps. This means you cannot use a multiplicative model for the day-of-the-week and an additive model for the day-of-the-year. Therefore, the global specification og log = TRUE or log = FALSE is used for all steps.
}
\examples{
Sys.setenv("JAVA_HOME"="C:/Workspace/Java/JDK/jdk-17.0.3+7") ## Currently start with this
.libPaths("C:/Workspace/R/JD_lib") ## Currently start with this
## Create time series 
set.seed(2358)
all <- tssim::sim_daily(N=5)
series <- all$original 

## Default adjustment
result <- dsa(series)
plot(result)

## Set STL parameters to be used
result2 <- dsa(series, s7 = stl_method(swindow = 31, twindow=9), pre_processing = result)

## Use STL and X11 in combination
result3 <- dsa(series, s7 = "x11", s365 = "stl", pre_processing = result) 

## Compare results
compare_plot(result2, result3)
}
\references{
Ollech, Daniel (2018). Seasonal Adjustment of Daily Time Series.
Bundesbank Discussion Paper 41/2018.

Ollech, Daniel (2021). Seasonal Adjustment of Daily Time Series.
Journal of Time Series Econometrics 13 (2), 235-264.
}
\author{
Daniel Ollech, Christiane Hofer, Martin Stefan, Thomas Witthohn
}
