% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{output}
\alias{output}
\title{Output for dsa2}
\usage{
output(dsa2_object)
}
\arguments{
\item{dsa2_object}{output object}
}
\description{
Output for dsa2
}
\details{
Generates a .Rdm file that is rendered into an html document saved in the working directory.
}
\author{
Lea Hengen, Sindy Brakemeier
}
