% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{output}
\alias{output}
\title{HTML output for dsa2}
\usage{
output(x, fileName = NULL, filePath = NULL)
}
\arguments{
\item{x}{output object}

\item{fileName}{name for HTML output}

\item{filePath}{path for HTML output}
}
\description{
HTML output for dsa2
}
\details{
Generates an .Rmd file that is rendered into an .html document saved in the working directory.
}
\author{
Lea Hengen, Sindy Brakemeier
}
