% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot.dsa2}
\alias{plot.dsa2}
\title{Plot generic for dsa2}
\usage{
\method{plot}{dsa2}(
  x,
  main = "Result for seasonal adjustment of daily time series",
  include_forecasts = FALSE,
  ...
)
}
\arguments{
\item{x}{dsa2-output object}

\item{main}{title of the plot}

\item{include_forecasts}{display forecast data}

\item{...}{additional parameters from plot() function}
}
\description{
Creates a plot of original and seasonally adjusted series.
}
\details{
The function uses the base plot package.
}
\author{
Sindy Brakemeier, Daniel Ollech
}
