% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot.dsa2}
\alias{plot.dsa2}
\title{Plot generic for dsa2}
\usage{
\method{plot}{dsa2}(
  x,
  main = "Result for seasonal adjustment of daily time series",
  include_forecasts = FALSE,
  type = "l",
  xlab = "",
  ylab = "",
  cex.axis = 0.75,
  bty = "n",
  mar = c(4, 2, 2, 0.5),
  ...
)
}
\arguments{
\item{x}{dsa2-output object}

\item{main}{title of the plot}

\item{include_forecasts}{display forecast data}

\item{type}{line type}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis, you may need to change the margins using mar}

\item{cex.axis}{cex.axis see ?par}

\item{bty}{bty see ?par}

\item{mar}{margins of the plot}

\item{...}{additional parameters from plot() function}
}
\description{
Creates a plot of original and seasonally adjusted series.
}
\details{
The function uses the base plot package.
}
\author{
Sindy Brakemeier, Daniel Ollech
}
