% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot_interactive.dsa2}
\alias{plot_interactive.dsa2}
\title{Interactive plot for dsa2}
\usage{
\method{plot_interactive}{dsa2}(x, ...)
}
\arguments{
\item{x}{dsa2-output object}

\item{...}{additional parameters for dygraphs::dyOptions}
}
\description{
Creates a plot of original and seasonally adjusted series.
}
\details{
The function uses the dygraphs package.
}
\author{
Martin Stefan, Daniel Ollech
}
