% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Generic for interactive plots}
\usage{
plot_interactive(x, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{parameters}
}
\description{
Generic for interactive plots
}
\author{
Daniel Ollech
}
