% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{plot_interactive.xts}
\alias{plot_interactive.xts}
\title{Interactive plot for xts}
\usage{
\method{plot_interactive}{xts}(x, ...)
}
\arguments{
\item{x}{xts time series}

\item{...}{additional parameters for dygraphs::dyOptions}
}
\description{
Creates a plot for xts time series
}
\details{
The function uses the dygraphs package.
}
\author{
Daniel Ollech
}
