% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{print.dsa2}
\alias{print.dsa2}
\title{Print generic for dsa2}
\usage{
\method{print}{dsa2}(x, ...)
}
\arguments{
\item{x}{dsa2-output object}

\item{...}{further arguments to print}
}
\description{
Print generic for dsa2, lists the coefficients of the fractional airline model as well as the coefficients and t-values of all outliers and calendar effects.
}
\author{
Sindy Brakemeier, Lea Hengen
}
