% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{seats_method}
\alias{seats_method}
\title{Handler for seats}
\usage{
seats_method(
  period = NA,
  log = FALSE,
  sn = FALSE,
  stde = FALSE,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{period}{period}

\item{log}{log, ignored in dsa2}

\item{sn}{sn}

\item{stde}{stde}

\item{nbcasts}{nbcasts}

\item{nfcasts}{nfcasts}
}
\description{
Handler for seats
}
\details{
This functions is basically a translator between the dsa2 routines
and rjd3highfreq::fractionalAirlineDecomposition, but its goal is to invoke the Seats procedure.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in DSA2.
}
\author{
Daniel Ollech
}
