% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{seats_method}
\alias{seats_method}
\title{Translation between input to dsa() and rjd3-function}
\usage{
seats_method(
  period = NA,
  log = NULL,
  sn = FALSE,
  stde = FALSE,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{period}{frequency which shall be adjusted, numeric}

\item{log}{log, ignored in dsa2}

\item{sn}{shall the series be decomposed into two components (signal and noise)}

\item{stde}{boolean, shall standard deviations of components be computed}

\item{nbcasts}{number of forecasts}

\item{nfcasts}{number of backcasts}
}
\description{
Handler for seats
}
\details{
This functions is basically a translator between the dsa2 routines
and rjd3highfreq::fractionalAirlineDecomposition, but its goal is to invoke the Seats procedure.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in DSA2.
}
\author{
Daniel Ollech
}
