% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{seats_method}
\alias{seats_method}
\title{Handler for seats}
\usage{
seats_method(
  period = NA,
  log = TRUE,
  sn = FALSE,
  stde = FALSE,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{period}{period}

\item{log}{log}

\item{sn}{sn}

\item{stde}{stde}

\item{nbcasts}{nbcasts}

\item{nfcasts}{nfcasts}
}
\description{
Handler for seats
}
\author{
Daniel Ollech
}
