% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{spectrum.dsa2}
\alias{spectrum.dsa2}
\title{Plot the periodogram of a daily time series}
\usage{
\method{spectrum}{dsa2}(dsa2_object, xlog = FALSE)
}
\arguments{
\item{dsa2_object}{dsa2-object}
}
\description{
Plot the periodogram of a daily time series
}
\details{
The spectrum is build around the spec.pgram() function
}
\examples{
x <- tssim::sim_daily(3)$original
res <- dsa(x)
spectrum(res)
}
\author{
Daniel Ollech
}
