% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{spectrum.numeric}
\alias{spectrum.numeric}
\title{Spectrum for numeric object}
\usage{
\method{spectrum}{numeric}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{parameters}
}
\description{
See ?stats::spectrum  for details
}
