% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{spectrum.ts}
\alias{spectrum.ts}
\title{Spectrum for ts object}
\usage{
\method{spectrum}{ts}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{parameters}
}
\description{
See ?stats::acf for details
}
\keyword{internal}
