% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{stats_stl_method}
\alias{stats_stl_method}
\title{Translation between input to dsa() and stats::stl-function}
\usage{
stats_stl_method(
  swindow = 13,
  twindow = NA,
  lwindow = NA,
  ninnerloop = 1,
  nouterloop = 15,
  sjump = NA,
  tjump = NA,
  ljump = NA
)
}
\arguments{
\item{swindow}{number of observations included in the local regressions calculated to obtain the seasonal component}

\item{twindow}{number of observations included in local regressions for trend component}

\item{lwindow}{number of observations included in local regressions for additional trend smoothing of seasonal}

\item{ninnerloop}{number of inner loops of STL}

\item{nouterloop}{number of outer loops of STL}

\item{sjump}{linear interpolation for seasonal component (see stats::stl)}

\item{tjump}{sjump linear interpolation for trend component (see stats::stl)}

\item{ljump}{linear interpolation for seasonal subseries component (see stats::stl)}
}
\description{
Handler for stats::stl
}
\details{
This functions is basically a translator between the dsa2 routines
and stats::stl.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in DSA2.
}
\author{
Daniel Ollech
}
