% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{stl_method}
\alias{stl_method}
\title{Handler for stl}
\usage{
stl_method(
  period = NA,
  swindow = 13,
  log = TRUE,
  twindow = 0,
  ninnerloop = 1,
  nouterloop = 15,
  nojump = FALSE,
  weight.threshold = 0.001,
  weight.function = c("BIWEIGHT")
)
}
\arguments{
\item{period}{period}

\item{swindow}{swindow}

\item{log}{log}

\item{twindow}{twindow}

\item{ninnerloop}{ninnerloop}

\item{nouterloop}{nouterloop}

\item{nojump}{nojump}

\item{weight.threshold}{weight.threshold}

\item{weight.function}{weight.function}
}
\description{
Handler for stl
}
\details{
This functions is basically a translator between the dsa2 routines
and rjd3stl::stl, but its goal is to invoke the stl procedure.
}
\author{
Daniel Ollech
}
