% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{stl_method}
\alias{stl_method}
\title{Translation between input to dsa() and rjd3-function}
\usage{
stl_method(
  period = NA,
  swindow = 13,
  log = NULL,
  twindow = 0,
  ninnerloop = 1,
  nouterloop = 15,
  nojump = FALSE,
  weight.threshold = 0.001,
  weight.function = c("BIWEIGHT")
)
}
\arguments{
\item{period}{frequency which shall be adjusted, numeric}

\item{swindow}{number of observations included in the local regressions calculated to obtain the seasonal component}

\item{log}{log, ignored in dsa2}

\item{twindow}{number of observations included in local regressions for trend component}

\item{ninnerloop}{number of inner loops of STL}

\item{nouterloop}{number of outer loops of STL}

\item{nojump}{impacts the precision of the estimation}

\item{weight.threshold}{threshold for weights, see ? rjd3stl::stlplus for details}

\item{weight.function}{wfunction for weights, see ? rjd3stl::stlplus for details}
}
\description{
Handler for stl
}
\details{
This functions is basically a translator between the dsa2 routines
and rjd3stl::stlplus, but its goal is to invoke the stl procedure.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in DSA2.
}
\author{
Daniel Ollech
}
