% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{summary.dsa2}
\alias{summary.dsa2}
\title{Summary generic for dsa2}
\usage{
\method{summary}{dsa2}()

\method{summary}{dsa2}()
}
\description{
Summary generic for dsa2

Summary generic for dsa2
}
\author{
x
}
