% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2_auxiliary.R
\name{summary.dsa2}
\alias{summary.dsa2}
\title{Summary generic for dsa2}
\usage{
\method{summary}{dsa2}(object, ...)
}
\arguments{
\item{object}{dsa2 output object}

\item{...}{further arguments handed to print.dsa2}
}
\description{
Summary.dsa2 lists the coefficients of the fractional airline model as well as the coefficients and t-values of all outliers and calendar effects.
}
\author{
Sindy Brakemeier, Lea Hengen
}
