% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{x11_method}
\alias{x11_method}
\title{Translation between input to dsa() and rjd3-function}
\usage{
x11_method(
  period = NA,
  log = NULL,
  sma = c("S3X9", "S3X1", "S3X3", "S3X5", "S3X15")[1],
  trend.horizon = 6,
  trend.degree = 2,
  trend.kernel = c("Henderson"),
  trend.asymmetric = c("CutAndNormalize"),
  sigma = c(1.5, 2.5)
)
}
\arguments{
\item{period}{frequency which shall be adjusted, numeric}

\item{log}{log, ignored in dsa2}

\item{sma}{seasonal moving average used}

\item{trend.horizon}{band.width of trend filters, see ?rjd3x11plus::x11plus for details}

\item{trend.degree}{polynomial order of the local trend model, see ?rjd3x11plus::x11plus for details}

\item{trend.kernel}{kernel weights in objective function, see ?rjd3x11plus::x11plus for details}

\item{trend.asymmetric}{truncation type to obtain asymmetric from symmetric filter, see ?rjd3x11plus::x11plus for details}

\item{sigma}{lower and upper limit for the extreme value detection}
}
\description{
Handler for x-11
}
\details{
This functions is basically a translator between the dsa2 routines
and rjd3x11plus::x11plus, but its goal is to invoke the X-11 procedure.
It cannot be used to change the decomposition scheme (additive/multiplicative) for a single step in DSA2.
}
\author{
Daniel Ollech
}
