% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa2.R
\name{x11_method}
\alias{x11_method}
\title{Handler for x-11}
\usage{
x11_method(
  period = NA,
  log = TRUE,
  sma = c("S3X9", "S3X1", "S3X3", "S3X5", "S3X15")[1],
  trend.horizon = 6,
  trend.degree = 2,
  trend.kernel = c("Henderson"),
  trend.asymmetric = c("CutAndNormalize"),
  sigma = c(1.5, 2.5)
)
}
\arguments{
\item{period}{period}

\item{log}{log}

\item{sma}{sma}

\item{trend.horizon}{t}

\item{trend.kernel}{t}

\item{trend.asymmetric}{t}

\item{sigma}{sigma}
}
\description{
Handler for x-11
}
\author{
Daniel Ollech
}
